implement more geomap schema fields
